intrinsic-test: simplify architecture constants
There doesn't need to be so many or other modules with the values.
